requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.3
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.4.7
whitenoise==5.2.0
